Refine hero profile layout
